UPDATE MS_PARSER.c / MS_PARSE_TYPE.C : ADD MS_EXPANSION.C :
Update parser and parse type with better management of dquotes;
Add ms_expansion.c with expansion WIP;
