+++ sources/parsing/ms_expansion.c
@@ -1,0 +1,41 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ms_expansion.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jcervoni <jcervoni@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/10 15:00:46 by jcervoni          #+#    #+#             */
+/*   Updated: 2022/05/11 13:14:51 by jcervoni         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+// void	ft_to_expand(t_arg *arg, t_env *env)
+// {
+// 	int	i;
+// 	int	len;
+	
+// 	i = 0;
+// 	len = ft_strlen(arg->content);
+// 	if (arg->token == TOKEN_DQUOTE)
+// 	{
+// 		while (arg->content[i])
+// 		{
+// 			if (arg->content[i] == '"')
+				
+// 		}
+// 	}
+// }
+
+// void	ft_dquote_expand(t_arg *arg, t_env *env, int end)
+// {
+// 	char	*temp;
+// 	char	*sub;
+// 	int		i;
+
+// 	i = 0;
+// 	if (end != 0)
+		
+// }
